Fixed more Clang warnings

+++ libraries/provideo_protocol/provideo_protocol_auto.c
@@ -378,7 +378,7 @@ static int get_aec_weights
                     cnt++;
 
                     // decrement data counter by bytes taken out
-                    unsigned int num_bytes_processed = (unsigned int)s + offset - (unsigned int)data;
+                    unsigned int num_bytes_processed = (unsigned int)(s + offset - data);
                     data_count -= num_bytes_processed;
 
                     // move out the processed command
@@ -417,7 +417,7 @@ static int get_aec_weights
             // timeout handling, if device does not send new pixel positions after
             // 1 second, assume table has been transmitted completely
             get_time_monotonic( &now );
-            int diff_ms = (now.tv_sec - start.tv_sec) * 1000 + (now.tv_nsec - start.tv_nsec) / 1000000;
+            int diff_ms = (int)((now.tv_sec - start.tv_sec) * 1000 + (now.tv_nsec - start.tv_nsec) / 1000000);
             loop = (diff_ms > 1000) ? 0 : 1;
         }
     }
